feat: BaseROS2API class
